pyproject.toml
[tool.poetry]
name = "opentelemetry-intro-gcf-python"
version = "0.1.0"
description = ""
authors = ["danish.amjad <danish.amjad@justeattakeaway.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
functions-framework = "==3.*"
opentelemetry-api = "1.24.0"
opentelemetry-exporter-gcp-monitoring = "1.6.0a0"
opentelemetry-exporter-gcp-trace = "1.6.0"
opentelemetry-exporter-otlp = "1.24.0"
opentelemetry-instrumentation = "0.45b0"
opentelemetry-resourcedetector-gcp = "1.6.0a0"
opentelemetry-sdk = "1.24.0"
opentelemetry-semantic-conventions = "0.45b0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
